FIR/UAST: commonize invalid UClass
