+++ plugins/uast-kotlin-fir/src/org/jetbrains/uast/kotlin/declarations/FirKotlinUClass.kt
@@ -37,5 +37,3 @@ class FirKotlinUClass(
 // TODO: FirKotlinUAnonymousClass or FirKotlinUAnonymousObject ?
 
 // TODO: FirKotlinScriptUClass
-
-// TODO: FirKotlinInvalidUClass
